Use pkg-cache on host system for packages
We have used the pacman pkg cache on the SD card to always assure we get the newest packages.
But as some have pointed out, this makes the installer redownload all the packages again when rerun after a failed installation attempt.
This change makes us create a new pkg-cache on the host system, which we then mount into the SD card's filesystem,
so it can be used by pacman. This would make the installer only download new and/or failed packages if rerun.
Fixes #15.

Signed-off-by: Dan Johansen <strit@manjaro.org>